+++ R/exp_stats.R
@@ -14,8 +14,38 @@
 #' # Expected values
 #'
 #' The `expected` argument is optional. If provided, this argument must
-#' be a character vector with values corresponding to columns in `.data`
+#' be a character vector with values corresponding to column names in `.data`
 #' containing expected experience. More than one expected basis can be provided.
+#'
+#' # Control variables
+#'
+#' The `control_vars` argument is optional. If provided, this argument must
+#' be `".none"` (more on this below) or a character vector with values
+#' corresponding to column names in `.data`. Control variables are used to
+#' estimate the impact of any grouping variables on observed experience
+#' *after accounting for* the impact of control variables.
+#'
+#' Mechanically, when values are passed to `control_vars`, a separate call
+#' is made to [exp_stats()] using the control variables as grouping variables.
+#' This is used to derive a new expected values basis called `control`, which is
+#' both added to `.data` and appended to the `expected` argument. In the final
+#' output, a column called `ae_control` shows the relative impact of any
+#' grouping variables after accounting for the control variables.
+#'
+#' **About `".none"`**: If `".none"` is passed to `control_vars`, a single
+#' aggregate termination rate is calculated for the entire data set and used to
+#' compute `control` and `ae_control`.
+#'
+#' The `control_distinct_max` argument places an upper limit on the number of
+#' unique values that a control variable is allowed to have. This limit exists
+#' to prevent an excessive number of groups on continuous or high-cardinality
+#' features.
+#'
+#' It should be noted that usage of control variables is a rough approximation
+#' and not a substitute for rigorous statistical models. The impact of control
+#' variables is calculated in isolation and does consider other features or
+#' possible confounding variables. As such, control variables are most useful
+#' for exploratory data analysis.
 #'
 #' # Credibility
 #'
@@ -69,6 +99,10 @@
 #' method
 #' @param conf_int If `TRUE`, the output will include confidence intervals
 #' around the observed termination rates and any actual-to-expected ratios.
+#' @param control_vars `".none"` or a character vector containing column names
+#' in `.data` to use as control variables
+#' @param control_distinct_max Maximum number of unique values allowed for
+#' control variables
 #' @param object An `exp_df` object
 #' @param ... Groups to retain after `summary()` is called
 #'
@@ -76,8 +110,9 @@
 #' and `data.frame`. The results include columns for any grouping variables,
 #' claims, exposures, and observed termination rates (`q_obs`).
 #'
-#' - If any values are passed to `expected`, expected termination rates and
-#' actual-to-expected ratios.
+#' - If any values are passed to `expected` or `control_vars`, additional
+#' columns are added for expected termination rates and actual-to-expected
+#' (A/E) ratios. A/E ratios are prefixed by `ae_`.
 #' - If `credibility` is set to `TRUE`, additional columns are added
 #' for partial credibility and credibility-weighted termination rates
 #' (assuming values are passed to `expected`). Credibility-weighted termination
@@ -100,7 +135,7 @@
 #' exp_res <- census_dat |>
 #'            expose("2019-12-31", target_status = "Surrender") |>
 #'            group_by(pol_yr, inc_guar) |>
-#'            exp_stats()
+#'            exp_stats(control_vars = "product")
 #'
 #' exp_res
 #' summary(exp_res)
@@ -115,7 +150,9 @@ exp_stats <- function(.data, target_status = attr(.data, "target_status"),
                       wt = NULL,
                       credibility = FALSE,
                       conf_level = 0.95, cred_r = 0.05,
-                      conf_int = FALSE) {
+                      conf_int = FALSE,
+                      control_vars,
+                      control_distinct_max = 25L) {
 
   .groups <- groups(.data)
   start_date <- attr(.data, "start_date")
@@ -123,12 +160,12 @@ exp_stats <- function(.data, target_status = attr(.data, "target_status"),
 
   if (is.null(target_status)) {
     target_status <- levels(.data$status)[-1]
-    rlang::warn(c(x = "No target status was provided.",
-                  i = glue::glue("{paste(target_status, collapse = ', ')} was assumed.")))
+    cli::cli_warn(c(x = "No target status was provided.",
+                    i = "{.val {target_status}} {?was/were} assumed."))
   }
 
   if (length(wt) > 1) {
-    rlang::abort(c(x = glue::glue("Only 1 column can be passed to `wt`. You supplied {length(wt)} values.")))
+    cli::cli_abort(c(x = "Only 1 column can be passed to `wt`. You supplied {length(wt)} values."))
   }
 
   check_split_expose_basis(.data, col_exposure)
@@ -151,31 +188,95 @@ exp_stats <- function(.data, target_status = attr(.data, "target_status"),
     res$claims <- res$n_claims
   }
 
+  if (missing(expected) || is.null(expected)) {
+    expected <- NULL
+  } else {
+    verify_col_exist(names(res), expected, "expected values column")
+  }
+
+  if (!missing(control_vars) && !is.null(control_vars)) {
+
+    # special handling for aggregates
+    if (all(control_vars == ".none")) {
+      if (".none" %in% names(res)) {
+        cli::cli_abort(
+          "Name conflict error: {.val .none} cannot be passed to `control_vars`
+          because there is also column in the data with the same name."
+        )
+      }
+      control_vars <- "None"
+      res[["None"]] <- 1L
+    } else {
+      if (".none" %in% control_vars) {
+        cli::cli_abort(
+          "If {.val .none} is passed to `control_vars`, then no other values
+          are allowed."
+        )
+      }
+      verify_col_exist(names(res), control_vars, "control variable")
+    }
+
+    # throw an error if too many unique values
+    nd_ctrl <- res |>
+      ungroup() |>
+      dplyr::summarize(dplyr::across(control_vars, dplyr::n_distinct))
+    nd_ctrl <- colnames(nd_ctrl)[nd_ctrl > control_distinct_max]
+    if (length(nd_ctrl) > 0) {
+      cli::cli_abort(c(
+        x = "There are too many distinct values in the {.val {nd_ctrl}} control variable{?s}.",
+        i = paste0(
+          "Limit = {control_distinct_max}. Update the `control_distinct_max` ",
+          "argument to increase the limit, or consider techniques like ",
+          "binning to reduce cardinality")
+      ))
+    }
+
+    # calculate observed rates across control variables
+    ctrl_dat <- finish_exp_stats(
+      res |> group_by(dplyr::across(dplyr::all_of(control_vars))),
+      target_status, expected = NULL, .groups = control_vars,
+      start_date = start_date, end_date = end_date,
+      credibility = FALSE, conf_level = conf_level,
+      cred_r = cred_r, wt = wt, conf_int = FALSE, control_vars = NULL
+    ) |>
+      select(dplyr::all_of(control_vars), control = q_obs)
+
+    # join observed rates on control variables to the data and add these
+    #   to expected values
+    res <- left_join(res, ctrl_dat, by = control_vars)
+    expected <- c(expected, "control")
+
+  } else {
+    control_vars <- NULL
+  }
+
   finish_exp_stats(res, target_status, expected, .groups,
                    start_date, end_date, credibility,
-                   conf_level, cred_r, wt, conf_int)
+                   conf_level, cred_r, wt, conf_int, control_vars)
 
 }
 
 #' @export
 print.exp_df <- function(x, ...) {
 
-  cat("Experience study results\n\n")
+  cli::cli_h2("Experience study results")
   if (length(groups(x)) > 0) {
-    cat(" Groups:", paste(groups(x), collapse = ", "), "\n")
+    cli::cli_ul("{.field Groups}: {groups(x)}")
   }
-  cat(" Target status:", paste(attr(x, "target_status"), collapse = ", "), "\n",
-      "Study range:", as.character(attr(x, "start_date")), "to",
-      as.character(attr(x, "end_date")), "\n")
+  cli::cli_ul(c(
+    "{.field Target status}: {attr(x, 'target_status')}",
+    "{.field Study range}: {attr(x, 'start_date')} to {attr(x, 'end_date')}"))
+  if (!is.null(attr(x, "control_vars"))) {
+    cli::cli_ul("{.field Control variables}: {attr(x, 'control_vars')}")
+  }
   if (!is.null(attr(x, "expected"))) {
-    cat(" Expected values:", paste(attr(x, "expected"), collapse = ", "), "\n")
+    cli::cli_ul("{.field Expected values}: {attr(x, 'expected')}")
   }
-  if (is.null(attr(x, "wt"))) {
-    cat("\n")
-  } else {
-    cat(" Weighted by:", attr(x, "wt"), "\n\n")
+  if (!is.null(attr(x, "wt"))) {
+    cli::cli_ul("{.field Weighted by}: {attr(x, 'wt')}\n")
   }
 
+  cat("\n")
   NextMethod()
 }
 
@@ -198,11 +299,12 @@ summary.exp_df <- function(object, ...) {
   expected <- attr(object, "expected")
   xp_params <- attr(object, "xp_params")
   wt <- attr(object, "wt")
+  control_vars <- attr(object, "control_vars")
 
   finish_exp_stats(res, target_status, expected, .groups,
                    start_date, end_date, xp_params$credibility,
                    xp_params$conf_level, xp_params$cred_r,
-                   wt, xp_params$conf_int)
+                   wt, xp_params$conf_int, control_vars)
 
 }
 
@@ -213,16 +315,16 @@ summary.exp_df <- function(object, ...) {
 finish_exp_stats <- function(.data, target_status, expected,
                              .groups, start_date, end_date,
                              credibility, conf_level, cred_r,
-                             wt, conf_int) {
+                             wt, conf_int, control_vars) {
 
   # expected value formulas. these are already weighted if applicable
-  if (!missing(expected)) {
+  if (!is.null(expected)) {
     ex_mean <- exp_form("weighted.mean({.col}, exposure)",
                         "{.col}", expected)
     ex_ae <- exp_form("q_obs / {.col}",
                       "ae_{.col}", expected)
   } else {
-    ex_ae <- ex_mean <- expected <- NULL
+    ex_ae <- ex_mean <- NULL
   }
 
   # additional columns for weighted studies
@@ -341,13 +443,14 @@ finish_exp_stats <- function(.data, target_status, expected,
              wt = wt,
              credibility = credibility,
              conf_level = conf_level, cred_r = cred_r,
-             conf_int = conf_int)
+             conf_int = conf_int,
+             control_vars = control_vars)
 }
 
 # low level class constructor
 new_exp_df <- function(x, .groups, target_status, start_date, expected,
                        end_date, wt, credibility, conf_level,
-                       cred_r = cred_r, conf_int) {
+                       cred_r = cred_r, conf_int, control_vars) {
   tibble::new_tibble(x,
                      class = "exp_df",
                      groups = .groups,
@@ -359,7 +462,8 @@ new_exp_df <- function(x, .groups, target_status, start_date, expected,
                      xp_params = list(credibility = credibility,
                                       conf_level = conf_level,
                                       cred_r = cred_r,
-                                      conf_int = conf_int))
+                                      conf_int = conf_int),
+                     control_vars = control_vars)
 }
 
 # this function is used to create formula specifications passed to dplyr::mutate
@@ -382,8 +486,8 @@ exp_form <- function(form, new_col, .col) {
 
 verify_exp_df <- function(.data) {
   if (!inherits(.data, "exp_df")) {
-    rlang::abort(c(x = glue::glue("`{deparse(substitute(.data))}` must be an `exp_df` object."),
-                   i = "Hint: Use `exp_stats()` to create `exp_df` objects."
+    cli::cli_abort(c(x = "`{deparse(substitute(.data))}` must be an `exp_df` object.",
+                     i = "Hint: Use `exp_stats()` to create `exp_df` objects."
     ))
   }
 }